Finally getting some output on the screen

+++ src/main/kotlin/modloader/ModInfo.kt
@@ -3,4 +3,4 @@ package asteroid4.modloader
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ModInfo(val modLoaderVersion: Short, val name: String, val version : String)
+data class ModInfo(val modLoaderVersion: Int, val name: String, val version : String)
